test(bridge): kill missed mutants in bridge_persistence and bridge_retrieval
- AbsenceEntry::normalize: test trimming and lowercasing (kills String::new()/xyzzy mutants)
- AbsenceEntry::fnv1a_hash: test hash distinctness (kills 0/1 returns and ^= -> |=/&= mutants)
- AbsenceEntry::merge_with: test None->Some assignment and lower-score-unchanged boundary
  (kills arm-deletion and > -> >= mutants)
- BridgeRetrieval::query: add empty-namespace test to pin both sides of || short-circuit
- BridgeRetrieval::compile_packet: test dedup (!seen), token-budget arithmetic, and
  confidence mean (kills delete-!, /->%, <=->>, +->, +=->=  mutants)
- Exclude src/bin/csm.rs from mutation (CLI I/O entry points untestable via --lib)
- Split inline test modules to bridge_retrieval_tests.rs / bridge_persistence_tests.rs
  to keep each file under the 500-LOC gate; bridge_persistence_tests gated on
  feature = "persistence" matching its subject module
