+++ src/bridge_persistence.rs
@@ -275,7 +275,7 @@ impl AbsenceEntry {
     }
 
     /// FNV-1a 64-bit hash (stable across Rust versions/platforms).
-    fn fnv1a_hash(bytes: &[u8]) -> u64 {
+    pub(crate) fn fnv1a_hash(bytes: &[u8]) -> u64 {
         const OFFSET_BASIS: u64 = 0xcbf2_9ce4_8422_2325;
         const PRIME: u64 = 0x0000_0100_0000_01b3;
         let mut hash = OFFSET_BASIS;
@@ -349,135 +349,5 @@ pub async fn persist_absence(
             store.upsert_absence(&entry).await?;
             Ok(entry)
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
-    use super::*;
-    use crate::semantic_bridge::CanonicalConcept;
-    use tempfile::NamedTempFile;
-
-    #[tokio::test]
-    async fn test_save_and_load_canonical_concept() {
-        let temp = NamedTempFile::new().unwrap();
-        let path = temp.path().to_str().unwrap();
-        let persistence = Persistence::new_local(path).await.unwrap();
-
-        let concept = CanonicalConcept::new("test-concept")
-            .with_label("label1")
-            .with_label("label2")
-            .with_related("related-concept");
-
-        persistence
-            .save_canonical_concept("_default", &concept)
-            .await
-            .unwrap();
-
-        let loaded = persistence
-            .load_canonical_concept("_default", "test-concept")
-            .await
-            .unwrap();
-        assert!(loaded.is_some());
-
-        let loaded = loaded.unwrap();
-        assert_eq!(loaded.id, "test-concept");
-        assert_eq!(loaded.labels, vec!["label1", "label2"]);
-        assert_eq!(loaded.related, vec!["related-concept"]);
-    }
-
-    #[tokio::test]
-    async fn test_delete_canonical_concept() {
-        let temp = NamedTempFile::new().unwrap();
-        let path = temp.path().to_str().unwrap();
-        let persistence = Persistence::new_local(path).await.unwrap();
-
-        let concept = CanonicalConcept::new("to-delete");
-        persistence
-            .save_canonical_concept("_default", &concept)
-            .await
-            .unwrap();
-
-        persistence
-            .delete_canonical_concept("_default", "to-delete")
-            .await
-            .unwrap();
-
-        let loaded = persistence
-            .load_canonical_concept("_default", "to-delete")
-            .await
-            .unwrap();
-        assert!(loaded.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_save_and_load_concept_graph() {
-        let temp = NamedTempFile::new().unwrap();
-        let path = temp.path().to_str().unwrap();
-        let persistence = Persistence::new_local(path).await.unwrap();
-
-        let mut graph = ConceptGraph::new();
-        graph.add_concept(
-            CanonicalConcept::new("c1")
-                .with_label("label1")
-                .with_related("c2"),
-        );
-        graph.add_concept(CanonicalConcept::new("c2").with_label("label2"));
-
-        persistence
-            .save_concept_graph("_default", &graph)
-            .await
-            .unwrap();
-
-        let loaded = persistence.load_concept_graph("_default").await.unwrap();
-        assert_eq!(loaded.concept_count(), 2);
-        assert_eq!(loaded.label_count(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_absence_id_generation() {
-        let q1 = "  Test Query  ";
-        let q2 = "test query";
-        let id1 = AbsenceEntry::id_for(q1);
-        let id2 = AbsenceEntry::id_for(q2);
-        assert_eq!(id1, id2);
-        assert!(id1.starts_with("absence:"));
-    }
-
-    #[tokio::test]
-    async fn test_persist_absence_lifecycle() {
-        let temp = NamedTempFile::new().unwrap();
-        let path = temp.path().to_str().unwrap();
-        let persistence = Persistence::new_local(path).await.unwrap();
-
-        let abstention = RetrievalAbstention {
-            query: "unknown".to_string(),
-            min_score_threshold: 0.5,
-            best_score_seen: Some(0.1),
-            attempted_modes: vec!["Auto".to_string()],
-            timestamp: Utc::now(),
-        };
-
-        // Create
-        let mut abstention = abstention;
-        abstention.best_score_seen = Some(0.1);
-        let entry = persist_absence(&abstention, &persistence).await.unwrap();
-        assert_eq!(entry.attempt_count, 1);
-        assert!((entry.best_score_ever.unwrap() - 0.1).abs() < f32::EPSILON);
-
-        // Update with higher score
-        let mut abstention2 = abstention.clone();
-        abstention2.best_score_seen = Some(0.4);
-        let entry2 = persist_absence(&abstention2, &persistence).await.unwrap();
-        assert_eq!(entry2.attempt_count, 2);
-        assert!((entry2.best_score_ever.unwrap() - 0.4).abs() < f32::EPSILON);
-
-        // Update with lower score
-        let mut abstention3 = abstention.clone();
-        abstention3.best_score_seen = Some(0.2);
-        let entry3 = persist_absence(&abstention3, &persistence).await.unwrap();
-        assert_eq!(entry3.attempt_count, 3);
-        assert!((entry2.best_score_ever.unwrap() - 0.4).abs() < f32::EPSILON); // Stays at 0.4
     }
 }
