+++ src/bridge_retrieval.rs
@@ -202,7 +202,7 @@ impl BridgeRetrieval {
     }
 
     /// Compute final score from breakdown using configurable weights.
-    fn compute_final_score(&self, scores: &ScoreBreakdown) -> f32 {
+    pub(crate) fn compute_final_score(&self, scores: &ScoreBreakdown) -> f32 {
         self.config.deterministic_weight * scores.deterministic
             + self.config.concept_weight * scores.concept
             + self.config.semantic_weight * scores.semantic
@@ -292,192 +292,5 @@ impl BridgeRetrieval {
     /// Get the configuration.
     pub const fn config(&self) -> &BridgeConfig {
         &self.config
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
-    use super::*; // Exact float comparisons for score test assertions
-    use crate::semantic_bridge::CanonicalConcept;
-    use crate::singularity::{Singularity, SingularityConfig};
-
-    #[test]
-    fn test_bridge_retrieval_empty_singularity() {
-        let encoder = TextEncoder::new();
-        let graph = ConceptGraph::new();
-        let bridge = BridgeRetrieval::with_defaults(encoder, graph);
-        let singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-
-        let results = bridge
-            .query("_default", &singularity, "test query", 10, None)
-            .unwrap();
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn test_bridge_retrieval_empty_graph() {
-        let encoder = TextEncoder::new();
-        let graph = ConceptGraph::new();
-        let bridge = BridgeRetrieval::with_defaults(encoder.clone(), graph);
-
-        let mut singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-        let concept = crate::singularity::ConceptBuilder::new("test-concept")
-            .with_vector(encoder.encode("test content"))
-            .build()
-            .unwrap();
-        singularity.inject("_default", concept).unwrap();
-
-        let results = bridge
-            .query("_default", &singularity, "test query", 10, None)
-            .unwrap();
-        // Should return deterministic results even without graph expansion
-        assert!(!results.is_empty());
-        assert!(results[0].scores.deterministic > 0.0);
-        assert!((results[0].scores.concept).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_bridge_retrieval_with_expansion() {
-        let encoder = TextEncoder::new();
-        let mut graph = ConceptGraph::new();
-
-        // Add canonical concept with label matching query
-        graph.add_concept(
-            CanonicalConcept::new("c1")
-                .with_label("agent-memory")
-                .with_label("session-context"),
-        );
-
-        let bridge = BridgeRetrieval::with_defaults(encoder.clone(), graph);
-
-        let mut singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-        let concept = crate::singularity::ConceptBuilder::new("mem-1")
-            .with_vector(encoder.encode("session context for AI agent"))
-            .build()
-            .unwrap();
-        singularity.inject("_default", concept).unwrap();
-
-        let results = bridge
-            .query("_default", &singularity, "agent memory session", 10, None)
-            .unwrap();
-
-        assert!(!results.is_empty());
-        // Check that expansion added concept score evidence
-        assert!(
-            results[0]
-                .scores
-                .evidence
-                .contains(&"deterministic_recall".to_string())
-        );
-    }
-
-    #[test]
-    fn test_memory_packet_empty_hits() {
-        let encoder = TextEncoder::new();
-        let graph = ConceptGraph::new();
-        let bridge = BridgeRetrieval::with_defaults(encoder, graph);
-        let singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-
-        let packet = bridge
-            .memory_packet("_default", &singularity, "test query", 10, None)
-            .unwrap();
-        assert!(packet.facts.is_empty());
-        assert!(packet.sources.is_empty());
-        assert!((packet.confidence).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_final_score_weights() {
-        let config = BridgeConfig {
-            deterministic_weight: 0.6,
-            concept_weight: 0.3,
-            semantic_weight: 0.1,
-            ..Default::default()
-        };
-
-        let encoder = TextEncoder::new();
-        let graph = ConceptGraph::new();
-        let bridge = BridgeRetrieval::new(encoder, graph, config);
-
-        let scores = ScoreBreakdown {
-            deterministic: 1.0,
-            concept: 1.0,
-            semantic: 1.0,
-            final_score: 0.0,
-            evidence: vec!["test".to_string()],
-        };
-
-        let final_score = bridge.compute_final_score(&scores);
-        assert!((final_score - 1.0).abs() < 1e-6); // All weights sum to 1.0
-    }
-
-    /// Kills compute_final_score -> 1.0 mutant: partial scores must NOT yield 1.0.
-    #[test]
-    fn test_final_score_partial_deterministic_only() {
-        let config = BridgeConfig {
-            deterministic_weight: 0.6,
-            concept_weight: 0.3,
-            semantic_weight: 0.1,
-            ..Default::default()
-        };
-        let bridge = BridgeRetrieval::new(TextEncoder::new(), ConceptGraph::new(), config);
-        let scores = ScoreBreakdown {
-            deterministic: 0.5,
-            concept: 0.0,
-            semantic: 0.0,
-            final_score: 0.0,
-            evidence: vec![],
-        };
-        // 0.6 * 0.5 = 0.3, not 1.0
-        let final_score = bridge.compute_final_score(&scores);
-        assert!(
-            (final_score - 0.3).abs() < 1e-5,
-            "expected 0.3, got {final_score}"
-        );
-    }
-
-    /// Kills query `||` -> `&&` mutant: top_k == 0 with non-empty singularity must return empty.
-    #[test]
-    fn test_query_top_k_zero_returns_empty() {
-        let encoder = TextEncoder::new();
-        let bridge = BridgeRetrieval::with_defaults(encoder.clone(), ConceptGraph::new());
-        let mut singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-        let concept = crate::singularity::ConceptBuilder::new("c1")
-            .with_vector(encoder.encode("hello world"))
-            .build()
-            .unwrap();
-        singularity.inject("_default", concept).unwrap();
-        // top_k == 0 must short-circuit even when singularity is non-empty
-        let results = bridge
-            .query("_default", &singularity, "hello", 0, None)
-            .unwrap();
-        assert!(results.is_empty(), "top_k=0 must yield empty results");
-    }
-}
-
-#[cfg(test)]
-mod tests_v2 {
-    #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
-    use super::*;
-    use crate::singularity::{ConceptBuilder, Singularity, SingularityConfig};
-    use csm_core_lib::hyperdim::HVec10240;
-
-    #[test]
-    fn test_bridge_retrieval_query_v2() {
-        let mut singularity = Singularity::<HVec10240>::new(SingularityConfig::default());
-        let concept = ConceptBuilder::new("c1")
-            .with_vector(HVec10240::random())
-            .build()
-            .unwrap();
-        singularity.inject("_default", concept).unwrap();
-
-        let bridge = BridgeRetrieval::new(
-            TextEncoder::new(),
-            ConceptGraph::new(),
-            BridgeConfig::default(),
-        );
-        let results = bridge.query("_default", &singularity, "test", 10, None);
-        assert!(results.is_ok());
     }
 }
